Service層をinterfaceと実装クラスに分割
